Many changes to make functions work together
I change how seats and main work, also had to do mini changes in many
functions to account for all further errors. It is building without
errors

+++ Economy.cpp
@@ -13,7 +13,7 @@ namespace Airline {
         CalculatePrice();
     }
 
-    int Economy::CalculatePrice() const{ // calculates the new price for the Economy object
+    double Economy::CalculatePrice() const{ // calculates the new price for the Economy object
 
         int days = getNumDays();
         double price = getPrice();
